+++ raven/client/util.py
@@ -91,10 +91,12 @@ def initialize_api(raven_server_url: str,
         api_config.raven_default_headers["Authorization"] = raven_api_key.strip()
 
     # Initialize audio player for TTS audio playback
-    if tts_playback_audio_device is not None:
+    if tts_playback_audio_device == "system-default":
+        logger.info("initialize_api: Using system's default TTS audio playback device. If you want to use another device, see `raven.client.config`, and run `raven-check-audio-devices` to get available choices.")
+    elif tts_playback_audio_device is not None:
         logger.info(f"initialize_api: Validating TTS audio playback device '{tts_playback_audio_device}' (from `raven.client.config`).")
     else:
-        logger.info("initialize_api: Using default TTS audio playback device. If you want to use a non-default device, see `raven.client.config`, and run `raven-check-audio-devices` to get available choices.")
+        logger.info("initialize_api: Using first available TTS audio playback device. If you want to use another device, see `raven.client.config`, and run `raven-check-audio-devices` to get available choices.")
     tts_playback_audio_device = audio_player.validate_playback_device(tts_playback_audio_device)
     api_config.audio_player = audio_player.Player(frequency=api_config.audio_frequency,
                                                   channels=2,
@@ -106,7 +108,7 @@ def initialize_api(raven_server_url: str,
     if stt_capture_audio_device is not None:
         logger.info(f"initialize_api: Validating STT audio capture device '{stt_capture_audio_device}' (from `raven.client.config`).")
     else:
-        logger.info("initialize_api: Using default STT audio capture device. If you want to use a non-default device, see `raven.client.config`, and run `raven-check-audio-devices` to get available choices.")
+        logger.info("initialize_api: Using first available STT audio capture device. If you want to use another device, see `raven.client.config`, and run `raven-check-audio-devices` to get available choices.")
     stt_capture_audio_device = audio_recorder.validate_capture_device(stt_capture_audio_device)
     api_config.audio_recorder = audio_recorder.Recorder(frame_length=512,  # TODO: need to be longer?
                                                         device_name=stt_capture_audio_device,